Changed the longs to ints becaues long is different on 32-bit and 64-bit processors.
git-svn-id: svn://svn.code.sf.net/p/wavbreaker/code/trunk@127 1b2899f7-b92e-0410-a9e2-f6ce2bc8120a

+++ src/wav.h
@@ -30,20 +30,20 @@ typedef char ID[4];
 
 typedef struct {
 	ID riffID;
-	unsigned long totSize;
+	unsigned int totSize;
 	ID wavID;
 } WaveHeader;
 
 typedef struct {
 	ID chunkID;
-	long chunkSize;
+	int chunkSize;
 } ChunkHeader;
 
 typedef struct {
 	short wFormatTag;
 	unsigned short  wChannels;
-	unsigned long   dwSamplesPerSec;
-	unsigned long   dwAvgBytesPerSec;
+	unsigned int   dwSamplesPerSec;
+	unsigned int   dwAvgBytesPerSec;
 	unsigned short  wBlockAlign;
 	unsigned short  wBitsPerSample;
 //	unsigned short  extraNonPcm;
